# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant Yemait EduCore school management API
servers:
  - url: /api
    description: Base API path
security:
  - ClerkAuth: []
...
paths:
...
  /academic/timetable:
    get:
      operationId: listAcademicTimetable
      tags: [academics]
      summary: List timetable entries in an authorized school
      parameters:
        - $ref: "#/components/parameters/SchoolId"
        - { name: sessionId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: termId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: classId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: section, in: query, schema: { type: string } }
        - { name: subjectId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: teacherId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: weekday, in: query, schema: { $ref: "#/components/schemas/Weekday" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/TimetableEntryStatus" } }
      responses:
        "200":
          description: Timetable entries
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AcademicTimetableEntry" } }
    post:
      operationId: createAcademicTimetableEntry
      tags: [academics]
      summary: Create a timetable entry without conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcademicTimetableEntryInput" }
      responses:
        "201":
          description: Timetable entry created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcademicTimetableEntry" }
  /academic/timetable/{entryId}:
    patch:
      operationId: updateAcademicTimetableEntry
      tags: [academics]
      summary: Update a timetable entry without conflicts
      parameters:
        - { name: entryId, in: path, required: true, schema: { type: integer, minimum: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcademicTimetableEntryUpdate" }
      responses:
        "200":
          description: Updated timetable entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcademicTimetableEntry" }
  /academic/students/me/timetable:
    get:
      operationId: getMyAcademicTimetable
      tags: [academics]
      summary: Get the authenticated student's class timetable
      parameters:
        - { name: schoolId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: sessionId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: termId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: weekday, in: query, schema: { $ref: "#/components/schemas/Weekday" } }
      responses:
        "200":
          description: Own class timetable
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AcademicTimetableEntry" } }
  /academic/parents/children/{studentId}/timetable:
    get:
      operationId: getChildAcademicTimetable
      tags: [academics]
      summary: Get a linked child's class timetable
      parameters:
        - $ref: "#/components/parameters/StudentIdPath"
        - { name: schoolId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: sessionId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: termId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: weekday, in: query, schema: { $ref: "#/components/schemas/Weekday" } }
      responses:
        "200":
          description: Linked child's class timetable
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AcademicTimetableEntry" } }
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: Clerk session token
      description: Clerk session cookie or bearer token validated by the API
    DeviceCredential:
      type: apiKey
      in: header
      name: X-Device-Credential
      description: Revocable device credential; the raw value is write-only and must never be logged or returned after rotation
...
  parameters:
    SchoolId:
      name: schoolId
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
...
    StudentIdPath:
      name: studentId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    TimetableEntryStatus:
      type: string
      enum: [ACTIVE, CANCELLED, ARCHIVED]
    Weekday:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
...
    AcademicTimetableEntry:
      type: object
      required: [id, schoolId, sessionId, termId, classId, section, subjectId, teacherId, weekday, startTime, endTime, room, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        schoolId: { type: integer }
        sessionId: { type: integer }
        termId: { type: integer }
        classId: { type: integer }
        section: { type: string }
        subjectId: { type: integer }
        teacherId: { type: integer }
        weekday: { $ref: "#/components/schemas/Weekday" }
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        room: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/TimetableEntryStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        className: { type: string }
        subjectName: { type: string }
        teacherFirstName: { type: string }
        teacherLastName: { type: string }
    AcademicTimetableEntryInput:
      type: object
      required: [schoolId, sessionId, termId, classId, subjectId, teacherId, weekday, startTime, endTime]
      properties:
        schoolId: { type: integer, minimum: 1 }
        sessionId: { type: integer, minimum: 1 }
        termId: { type: integer, minimum: 1 }
        classId: { type: integer, minimum: 1 }
        section: { type: string, maxLength: 80 }
        subjectId: { type: integer, minimum: 1 }
        teacherId: { type: integer, minimum: 1 }
        weekday: { $ref: "#/components/schemas/Weekday" }
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        room: { type: ["string", "null"], maxLength: 120 }
    AcademicTimetableEntryUpdate:
      type: object
      required: [schoolId]
      properties:
        schoolId: { type: integer, minimum: 1 }
        sessionId: { type: integer, minimum: 1 }
        termId: { type: integer, minimum: 1 }
        classId: { type: integer, minimum: 1 }
        section: { type: string, maxLength: 80 }
        subjectId: { type: integer, minimum: 1 }
        teacherId: { type: integer, minimum: 1 }
        weekday: { $ref: "#/components/schemas/Weekday" }
        startTime: { type: string, format: time }
        endTime: { type: string, format: time }
        room: { type: ["string", "null"], maxLength: 120 }
        status: { $ref: "#/components/schemas/TimetableEntryStatus" }
...